qxl: abstract object allocation away from surface/image/cursor code
This abstracts the object bo/surface allocation away from the user code.

The idea is we can then plug in a KMS backend for this abstraction and mostly
keep the surrounding code intact.

This is probably the biggest change to the driver in terms of impact of KMS
code on UMS code. At the moment I'm storing linked list of bos for release
handling, and that might be better done with a different data structure,
since we are looking them up by phy_addr, either a hash table or some sort
of rb tree perhaps, rather than a linear search.

This commit also starts usign the xorg list macros, which may require
some compat code to work in history.

Signed-off-by: Dave Airlie <airlied@redhat.com>