Implement sending audio to the client from a directory of FIFO queues
This introduces a new Xorg.conf option, SpicePlaybackFIFODir, which will
be monitored for files. The XSpice driver will mix and forward the audio
data sent to those pipes to the Spice client.

This is designed to work with PulseAudio's module-pipe-sink, but should
work with any audio output to a pipe. For example, use with this PA
configuration option:

  load-module module-pipe-sink file=$FIFO_DIR/playback.fifo format=s16 rate=44100 channels=2

making sure the format, rate, and channels match the Spice protocol
settings.