More correctly signal that we only want the first head connected at start.
This prevents a bug when using old versions of qemu (< 1.2) and the latest
driver version.

This was uncovered when I found that my patch beccf8e8:
  Establish a preferred default of 1024x768 correctly.
causes a failure when running with the qemu in Debian.

The video ram allocated by older qemu versions (8M) is not enough to hold the
default heads (4) times the default screen size (1024x768).

It worked prior to my patch because the previous code was a bit of a mess.
It would hard code 1024x768 for Virtual Size, hack some crtcs to disabled,
and sorta kinda have 4 heads connected. Later resizes would come through and
mostly fix things up, so we largely got away with.

This change should correctly signal that we start with only one head
connected.

There is still more that could be done with the randr code; we have
unimplemented functions that I suspect would allow us to fail more
graciously when the user tries to create displays in excess of our RAM.