xspice: add vdagent support
Adds a configurable virtio path used to communicate with the vdagentd,
and a configuration variable for enabling the thing.

With this you can have multiple monitors, but due to usage of a tablet
you cannot generate pointer events on and monitors besides the first.

clipboard already works.

The next patch adds uinput emulation to let vdagentd generate uinput
events and fix this glitch.

Signed-off-by: Alon Levy <alevy@redhat.com>