Remove image cache
While looking for leaks, I realized that the image cache looks
quite suspicious.

Not only it leaks when qxl_drop_image_cache() is called,
since all the allocated image_info_t references are lost.
But it is also useless:  "The name implies it's a cache of some
sort, but data is only added to it. The only time data is looked up in
that cache is when it needs to be destroyed, and this is not done
correctly and causes leaks. Since the cached data is never used and
since it's leaky, it's better to remove it altogether"