Use pci_io_write8 instead of outb
This ensures better portability as some arches (eg aarch64) don't have
outb.
On linux pci_io_* try to open the sysfs map file corresponding to the
I/O port range, which the kernel translates arch-appropriately.  If
there is no such file then it'll try to use port instructions if that's
a thing the architecture has.  So you only need one path, pciaccess
exists to get the portability right for you.

https://bugzilla.redhat.com/show_bug.cgi?id=1201877