spiceqxl_audio: Fix a race condition in the audio playback
can_feed() depends on the time and thus could return false in
process_fifos(), causing it to stop reading from the fifos, and then
true in watch_or_wait() so that the wall_timer would not be set, but
the fifos would not be watched either because they already contain
data to process. The audio playback would then come to a stop.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>