spiceqxl_audio: Only start the playback channel when fifos are present
This lets the client free the audio resources when they are not needed.
This matches QEMU's behavior.
As a side benefit it stops the client's mm-time from being stuck (due
to the audio backend's delay updates being applied to the mm-time of
the last audio message, that is the channel's creation) when no audio
application is running.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>