spiceqxl_audio: Stop the playback channel if there is nothing to play
This lets the client free the audio resources when an audio application
is not actually playing anything, typically because playback is paused.
This matches QEMU's behavior.
As a side benefit it stops the client's mm-time from being stuck (due
to the audio backend's delay updates being applied to the mm-time of
the last audio message) which lets video streams play in this situation.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>