xspice: Adjust to X.org 1.19 changes
In newer X.org versions, it's no longer supported to modify the set of
FDs passed to a BlockHandler method to get notified when the FD has data
to be read. This was limited anyway as we could only get read events
this way, and had to do our own polling to get notified about socket
writeability.

Starting from xserver 1.19, the supported way of doing this is to use
the SetNotifyFd/RemoveNotifyFd API, which is actually a much better way
as it matches very well the 'watch' API spice-server expects Xspice to
implement.

This commit switches to that new API, which removes the need for
RegisterBlockAndWakeupHandlers().

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Acked-by: Uri Lublin <uril@redhat.com>